+++ src/main/java/pl/asie/computronics/cc/ParticleTurtleUpgrade.java
@@ -7,6 +7,7 @@ import dan200.computercraft.api.peripheral.IComputerAccess;
 import dan200.computercraft.api.peripheral.IPeripheral;
 import dan200.computercraft.api.turtle.ITurtleAccess;
 import dan200.computercraft.api.turtle.TurtleSide;
+import net.minecraft.init.Blocks;
 import net.minecraft.init.Items;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.IIcon;
@@ -27,46 +28,50 @@ public class ParticleTurtleUpgrade extends TurtleUpgradeBase {
 		}
 
 		@Override
-		@Optional.Method(modid= Mods.ComputerCraft)
+		@Optional.Method(modid = Mods.ComputerCraft)
 		public String[] getMethodNames() {
-			return new String[]{"spawn"};
+			return new String[] { "spawn" };
 		}
 
 		@Override
-		@Optional.Method(modid=Mods.ComputerCraft)
+		@Optional.Method(modid = Mods.ComputerCraft)
 		public Object[] callMethod(IComputerAccess computer,
-				ILuaContext context, int method, Object[] arguments)
-				throws LuaException, InterruptedException {
+			ILuaContext context, int method, Object[] arguments)
+			throws LuaException, InterruptedException {
 			// check argument type validity
-			if(arguments.length < 4 || !(arguments[0] instanceof String)) return new Object[]{false, "invalid arguments"};
+			if(arguments.length < 4 || !(arguments[0] instanceof String))
+				return new Object[] { false, "invalid arguments" };
 			for(int i = 1; i < arguments.length; i++)
-				if(!(arguments[i] instanceof Double)) return new Object[]{false, "invalid argument "+i};
-			
-	        String name = (String)arguments[0];
+				if(!(arguments[i] instanceof Double))
+					return new Object[] { false, "invalid argument " + i };
 
-	        if (name.length() > Short.MAX_VALUE) {
-	            return new Object[]{false, "name too long"};
-	        }
+			String name = (String) arguments[0];
 
-	        Random rng = access.getWorld().rand;
-	        double x = access.getPosition().posX + 0.5 + (Double) arguments[1];
-	        double y = access.getPosition().posY + 0.5 + (Double) arguments[2];
-	        double z = access.getPosition().posZ + 0.5 + (Double) arguments[3];
-	        double defaultv = (rng.nextDouble() * 0.1);
-	        if(arguments.length >= 5) defaultv = (Double) arguments[4];
-	        double vx = defaultv * rng.nextGaussian();
-	        double vy = defaultv * rng.nextGaussian();
-	        double vz = defaultv * rng.nextGaussian();
-	        if(arguments.length >= 7) {
-	        	vx = (Double) arguments[4];
-	        	vy = (Double) arguments[5];
-	        	vz = (Double) arguments[6];
-	        }
-	        ParticleUtils.sendParticlePacket(name, access.getWorld(), x, y, z, vx, vy, vz);
-	        return new Object[]{true};
+			if(name.length() > Short.MAX_VALUE) {
+				return new Object[] { false, "name too long" };
+			}
+
+			Random rng = access.getWorld().rand;
+			double x = access.getPosition().posX + 0.5 + (Double) arguments[1];
+			double y = access.getPosition().posY + 0.5 + (Double) arguments[2];
+			double z = access.getPosition().posZ + 0.5 + (Double) arguments[3];
+			double defaultv = (rng.nextDouble() * 0.1);
+			if(arguments.length >= 5)
+				defaultv = (Double) arguments[4];
+			double vx = defaultv * rng.nextGaussian();
+			double vy = defaultv * rng.nextGaussian();
+			double vz = defaultv * rng.nextGaussian();
+			if(arguments.length >= 7) {
+				vx = (Double) arguments[4];
+				vy = (Double) arguments[5];
+				vz = (Double) arguments[6];
+			}
+			ParticleUtils.sendParticlePacket(name, access.getWorld(), x, y, z, vx, vy, vz);
+			return new Object[] { true };
 		}
-		
+
 	}
+
 	public ParticleTurtleUpgrade(int id) {
 		super(id);
 	}
@@ -78,7 +83,7 @@ public class ParticleTurtleUpgrade extends TurtleUpgradeBase {
 
 	@Override
 	public ItemStack getCraftingItem() {
-		return new ItemStack(Items.firework_charge, 1, 0);
+		return new ItemStack(Items.blaze_powder, 1, 0);
 	}
 
 	@Override
@@ -88,6 +93,10 @@ public class ParticleTurtleUpgrade extends TurtleUpgradeBase {
 
 	@Override
 	public IIcon getIcon(ITurtleAccess turtle, TurtleSide side) {
-		return Items.fireworks.getIconFromDamage(0);
+		if(turtle != null) {
+			int dyeColour = 15 - turtle.getDyeColour();
+			return dyeColour >= 0 && dyeColour < 16 ? Blocks.stained_glass.getIcon(0, dyeColour) : Blocks.glass.getIcon(0, 0);
+		}
+		return Blocks.glass.getIcon(0, 0);
 	}
 }